+++ src/prime_rl/entrypoints/sft.py
@@ -11,7 +11,7 @@ import tomli_w
 from prime_rl.configs.sft import SFTConfig
 from prime_rl.utils.config import cli
 from prime_rl.utils.logger import setup_logger
-from prime_rl.utils.pathing import get_config_dir, get_log_dir, validate_output_dir
+from prime_rl.utils.pathing import format_log_message, get_config_dir, get_log_dir, validate_output_dir
 from prime_rl.utils.process import cleanup_processes, cleanup_threads, monitor_process
 from prime_rl.utils.utils import get_free_port
 
@@ -79,14 +79,7 @@ def sft_slurm(config: SFTConfig):
 
     log_dir = get_log_dir(config.output_dir)
     num_nodes = config.deployment.num_nodes if config.deployment.type == "multi_node" else 1
-    col = 18
-    i1 = " " * 2
-    i2 = " " * 3
-    log_lines = [f"{i1}{'Trainer:':<{col}}tail -F {log_dir}/trainer.log"]
-    if num_nodes > 1:
-        log_lines.append(f"{i2}{'Nodes:':<{col - 1}}tail -F {log_dir}/trainer/node_*.log")
-    log_lines.append(f"{i2}{'Per-rank:':<{col - 1}}{log_dir}/trainer/torchrun/")
-    log_message = "Logs:\n" + "\n".join(log_lines)
+    log_message = format_log_message(log_dir=log_dir, trainer=True, num_train_nodes=num_nodes)
 
     if config.dry_run:
         logger.success(f"Dry run complete. To submit manually:\n\n  sbatch {script_path}\n\n{log_message}")
